Jocke solved the problem

+++ Solution/Maze.cpp
@@ -109,8 +109,8 @@ glm::vec3 Maze::TransformToMazeCoords(glm::vec3 pos)
 	newX += (this->GetMazeWidth() / 2) + pixelOffset;
 	newZ += (this->GetMazeHeight() / 2) + pixelOffset;
 
-	pos.x = (int)newX;
-	pos.z = (int)newZ;
+	pos.x = newX;
+	pos.z = newZ;
 
 	return pos;
 }
